requirements.txt
anyio==3.7.0
asgiref==3.5.0
bcrypt==4.0.1
blinker==1.4
certifi==2023.5.7
click==8.1.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info===1.1build1
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post2
exceptiongroup==1.1.1
fastapi==0.95.2
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.2
httplib2==0.20.2
httptools==0.5.0
httpx==0.24.1
idna==3.4
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.1.2
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
orjson==3.8.14
passlib==1.7.4
psycopg2==2.9.6
pyasn1==0.5.0
pydantic==1.10.8
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu1
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==5.4.1
rsa==4.9
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.15
starlette==0.27.0
systemd-python==234
typing_extensions==4.6.2
ubuntu-advantage-tools==8001
ufw==0.36.1
ujson==5.7.0
unattended-upgrades==0.1
uvicorn==0.22.0
uvloop==0.17.0
wadllib==1.3.6
watchfiles==0.19.0
websockets==11.0.3
wsproto==1.0.0
zipp==1.0.0
